backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pypdf==3.17.4
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
faiss-cpu==1.7.4
firebase-admin==6.2.0
python-dotenv==1.0.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0
pytest-mock==3.12.0 